refactor(yak): extract yakscript package and delegate ssa_compile/yakgrpc execution
Move plugin execution (ExecScriptWithParam, stream/feedback helpers) into
common/yak/yakscript; ssa_compile and yakgrpc debug/smoking paths become thin
forwarders. No behavior change beyond import graph and package layout.
